Fixed _frames folder generating in the wrong location
In the event that the path of a file to be converted contained special characters, these special
characters would be removed by the method to filter them out of the file name. The implementation
of this method has now been adjusted to only affect the file name instead of the whole directory.